+++ httpx/_utils.py
@@ -11,8 +11,6 @@ import typing
 from pathlib import Path
 from urllib.request import getproxies
 
-import sniffio
-
 from ._types import PrimitiveData
 
 if typing.TYPE_CHECKING:  # pragma: no cover
@@ -289,29 +287,18 @@ def peek_filelike_length(stream: typing.Any) -> int | None:
 
 
 class Timer:
-    async def _get_time(self) -> float:
-        library = sniffio.current_async_library()
-        if library == "trio":
-            import trio
-
-            return trio.current_time()
-        else:
-            import asyncio
-
-            return asyncio.get_event_loop().time()
-
     def sync_start(self) -> None:
         self.started = time.perf_counter()
 
     async def async_start(self) -> None:
-        self.started = await self._get_time()
+        self.started = time.perf_counter()
 
     def sync_elapsed(self) -> float:
         now = time.perf_counter()
         return now - self.started
 
     async def async_elapsed(self) -> float:
-        now = await self._get_time()
+        now = time.perf_counter()
         return now - self.started